% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARL-package.R
\docType{package}
\name{FARL-package}
\alias{FARL}
\alias{FARL-package}
\title{FARL: Factor-Augmented Regularized Latent Regression for Large-Scale Assessment}
\description{
FARL is developed to support large-scale assessment (LSA) analyses, in which
latent regression models are used to integrate students’ background information
and to generate plausible values (PVs) for secondary analysis. In LSA settings,
background variables are often high-dimensional and highly correlated, posing
substantial challenges for traditional latent regression approaches.
}
\details{
FARL implements factor-augmented regularized latent regression (FARLR), an
innovative framework that jointly models common factors and idiosyncratic
components. Regularization is employed to select relevant idiosyncratic
predictors, yielding interpretable regression results while maintaining
congeniality and estimation stability.
}
\section{Latent regression models}{

\itemize{
  \item \code{\link{Farlr_mml}} fits the factor-augmented regularized latent regression model via marginal maximum likelihood
    \itemize{
      \item Method 1: \code{FARLR_EMM}
      \item Method 2: \code{FARLR_Debias}
    }
}


\itemize{
 \item \code{\link{Dire_mml}} fits the DIRE latent regression model
}
}

\section{Plausible value generation}{

\itemize{
  \item \code{\link{Farlr_drawPVs}} generates plausible values under the FARLR framework
}


\itemize{
  \item \code{\link{Dire_drawPVs}} generates plausible values under the Dire framework
}
}

\section{Simulation and example data}{

\itemize{
  \item Built-in example datasets illustrating FARLR estimation and PV generation
  \item Utility functions for simulation studies in large-scale assessment settings
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://yijunchenguw.github.io/FARL/}
  \item \url{https://github.com/yijunchenguw/FARL}
}

}
\author{
\strong{Maintainer}: Yijun Cheng \email{chengxb@uw.edu} (\href{https://orcid.org/0000-0002-0671-9193}{ORCID})

}
\keyword{internal}
